Move all monitor classes to a monitor subdirectory
Problem:
Group all monitor code into resource/ltm/monitor

Analysis:
Group all monitor code into resource/ltm/monitor

Tests:
f5_cccl/resource/ltm/monitor/
