+++ f5_cccl/resource/ltm/monitor/test/test_tcp_monitor.py
@@ -5,7 +5,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    https://www.apache.org/licenses/LICENSE-2.0
+#    http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -15,13 +15,13 @@
 #
 
 import conftest
-import f5_cccl.resource.ltm.https_monitor as target
+import f5_cccl.resource.ltm.monitor.tcp_monitor as target
 
 
-class Test_HTTPSMonitor(conftest.TestLtmResource):
+class Test_TCPMonitor(conftest.TestLtmResource):
     pass  # any further deviation should be tested here...
 
 
 def test_entry():
-    assert target.HTTPSMonitor.monitor_schema_kvps._asdict() \
-        == target.default_schema, "Verified entry vector assignment"
+    assert target.TCPMonitor.monitor_schema_kvps._asdict() == \
+        target.default_schema, "Verified entry vector assignment"
